# backend/api-server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SLIME Tools External API",
    "version": "1.0.0",
    "description": "External API for third-party project integrations. All endpoints are scoped to programs created by the API key holder. Contact SLIME Tools to request an API key."
  },
  "servers": [
    { "url": "https://api.slime.tools/api/v1/external" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key"
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/staking-programs/{id}/register": {
      "post": {
        "summary": "Register a wallet and trigger immediate drip",
        "operationId": "registerParticipant",
        "description": "Registers a wallet and triggers an immediate reward drip in one of your programs.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "accountId": { "type": "string", "example": "0.0.12345" }
                },
                "required": ["accountId"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Registration result" },
          "400": { "description": "Not eligible or already registered" },
          "403": { "description": "Not authorized for this program" },
          "404": { "description": "Program not found" }
        }
      }
    },
...
  }
}